Fixed bug 130263: Polymorphing into a new version of yourself seems buggy.
The problem was that while newhp() and newexplevel() have been changed
to calculate HPs and energy levels based on the new NetHack system,
newman() was still using the old Slash'EM system. newexp() was different
from vanilla in two ways. Firstly it was compatible with the modified
newhp() and newexplevel() and secondly it generated new stats with no
regard to the current ones.

This had two major effects:
 i) Your stats could change by a much larger amount due to a freaky polymorph
 ii) Your stats were always in line with your new experience level.

It would have been possible to change newexp() so that it generated new
stats based soley on your experience level as before, but taking into
account the new methods of calculating HP and energy.

Instead I have changed newexp() to match vanilla in both ways. I did this
partly because it was simpler, and partly because I did not feel there was
any compelling reason to stick with the old method.