Implemented & tested the support routines for NhExt sub-protocol 1.
I've also thrown together an implementation of a server module
(which might be linked with a window interface to produce a
proxy server). There's no way it will compile yet.

The next stage is to try and see how this could dovetail into an
existing windowing interface. I really want to use GTK for this
since it is suitably complicated that it is likely to throw up
many difficulties and because it is the most important interface
to supply as a proxy module to solve our security problems.

Unfortunately, the GTK interface is very bad at doing the job it is
supposed to in the procedures. An example of this is select_player.
The interface currently grabs the values of initrole etc., before
the procedure is even called. It will take quite a while to iron
out these kind of problems before we can proceed.