Implemented first cut of display layers.
This extends the display module's concept of the hero's memory to be
layered. Currently, the layers are: background, traps, objects and
an invisible monster (normal monsters and special effects are not
remembered). When enabled (which it is by default in the devel
branch), the file format changes so games are not portable between
systems where one has this enabled and one disabled.

I have also added some layered support to the GTK2 port in preparation
for using this but there is currently no mechanism to expose the
display module's layered memory to the window port. This will be
added in due course.