Fixed bug 805499: Walking dead.
The problem that Nick describes stems from a confusion as to how the
amulet versus stoning works. It is (was) implemented as if it kicked
in before the hero died (which means that the number of deaths counter
was not increased and the amulet had no effect on the hero's HP, hunger,
sickness etc). This would be fine except that it was inconsistent in that
the player had already been told the hero was dead (either by an explicit
"You die" or by an implicit "You turned to stone"). Rather than try and
fix the messages, I have changed the amulet slightly so that it behaves
much more like the amulet of lifesaving. This means that the death is
counted and that HP etc will be affected (which solves the particular
problem that Nick encountered). This makes the amulet versus stoning
slightly more powerful than it used to be, but I don't think this is
significant.