Various bug fixes, courtesy of various people (mostly amy) by way of K2 (https://github.com/k21971/slashem/commits/hardfought): - Fix calm steed technique - Fix crash bug in tty when an item's description was just less than the window width, but description + accelerator + glyph + padding was just more than said width - Acid dropping fix (vanilla) - Deallocated wielded objects fix (vanilla) - Torch panic fix - Water punishment crash fix
There is another fix in k2's repo, which handles null pointer returns from makemon()
I am not merging that directly, because I would like to take a different approach to fixing it: the caller should check /if it's possible/ to create a monster first, and makemon() should panic if it can't make a mon.
This is not a terribly complicated change, but I want to avoid nontrivial changes before makedefs is gone, because that will just make my life so much easier.  Really, I'm not sure I should be doing anything important before the level/dungeon compilers are gone.