You are never "teetering on the edge of a pit" if you can fly
In particular, this means you can freely interact with items at the
bottom of a pit, solving the "a leprechaun steals all of your money,
then falls into a pit and dies, rendering the money inaccessible if you
are a vampire" problem.

It also means that if you drop items while flying over a pit, you get
the normal drop message rather than the "item falls into a pit" message;
one can assume that (as with picking items up) the player swoops down to
the bottom in a controlled manner to drop the items.